added alerts for deleting

+++ src/main/frontend/app/src/component/EditQuestionComponent.js
@@ -21,7 +21,8 @@ class AnswerRow extends Component {
         }
 
         function deleteAnswer(id) {
-            AnswerService.deleteAnswers(id);
+            let confirmed = window.confirm("Jeste li sigurni da želite izbrisati odgovor?");
+            if(confirmed) AnswerService.deleteAnswers(id);
             window.location.reload(false);
         }
